Extract mock_engine for tests
